+++ Products/CMFPlone/tests/testResourceRegistries.py
@@ -11,8 +11,10 @@ from Products.CMFPlone.interfaces import IResourceRegistry
 from Products.CMFPlone.interfaces.resources import OVERRIDE_RESOURCE_DIRECTORY_NAME
 from Products.CMFPlone.resources import add_bundle_on_request
 from Products.CMFPlone.resources import add_resource_on_request
+from Products.CMFPlone.resources import remove_bundle_on_request
 from Products.CMFPlone.resources.browser.cook import cookWhenChangingSettings
 from Products.CMFPlone.resources.browser.scripts import ScriptsView
+from Products.CMFPlone.resources.browser.styles import StylesView
 from Products.CMFPlone.resources.exportimport.resourceregistry import ResourceRegistryNodeAdapter
 from Products.CMFPlone.tests import PloneTestCase
 from Products.GenericSetup.context import SetupEnviron
@@ -301,9 +303,23 @@ class TestControlPanel(PloneTestCase.PloneTestCase):
 
 
 class DummyResource(object):
+    def __init__(self, name):
+        self.js = name
+        self.css = [name, ]
 
-    def __init__(self, js):
-        self.js = js
+
+class DummyBundle(object):
+    def __init__(self, name, enabled=True):
+        self.__prefix__ = 'test/' + name
+        self.compile = True
+        self.conditionalcomment = None
+        self.csscompilation = '++resource++' + name + '.css'
+        self.depends = None
+        self.enabled = enabled
+        self.expression = None
+        self.jscompilation = '++resource++' + name + '.js'
+        self.last_compilation = '123'
+        self.resources = []
 
 
 class TestScriptsViewlet(PloneTestCase.PloneTestCase):
@@ -359,3 +375,135 @@ class TestScriptsViewlet(PloneTestCase.PloneTestCase):
         add_bundle_on_request(req, 'foo')
 
         self.assertEqual(len(req.enabled_bundles), 1)
+
+    @mock.patch.object(
+        ScriptsView,
+        'get_bundles',
+        new=lambda self: {'foo': DummyBundle('foo', enabled=False)}
+    )
+    def test_add_bundle_on_request_with_subrequest(self):
+        req = self.layer['request']
+
+        # create a subrequest.
+        subreq = req.clone()
+        subreq['PARENT_REQUEST'] = req
+
+        # add a bundle via the main request
+        add_bundle_on_request(req, 'foo')
+
+        scripts = ScriptsView(self.layer['portal'], subreq, None)
+
+        # Send resource registry in development mode
+        # Via a fake registry to allow accessing like this:
+        # self.registry.records['plone.resources.development'].value
+        scripts.registry = type(
+            'reg',
+            (object, ),
+            {'records': {
+                'plone.resources.development': type(
+                    'val',
+                    (object, ),
+                    {'value': True}
+                )()
+            }}
+        )()
+        self.assertTrue(scripts.development)
+
+        scripts.update()
+        results = scripts.scripts()
+        self.assertEqual(
+            results[-1],
+            {
+                'src': 'http://nohost/plone/++resource++foo.js?version=123',
+                'conditionalcomment': None,
+                'bundle': 'foo'
+            }
+        )
+
+    @mock.patch.object(
+        ScriptsView,
+        'get_bundles',
+        new=lambda self: {'foo': DummyBundle('foo', enabled=True)}
+    )
+    def test_remove_bundle_on_request_with_subrequest(self):
+        req = self.layer['request']
+
+        # create a subrequest.
+        subreq = req.clone()
+        subreq['PARENT_REQUEST'] = req
+
+        # remove the enabled 'foo' bundle
+        remove_bundle_on_request(req, 'foo')
+
+        scripts = ScriptsView(self.layer['portal'], subreq, None)
+
+        # Send resource registry in development mode
+        # Via a fake registry to allow accessing like this:
+        # self.registry.records['plone.resources.development'].value
+        scripts.registry = type(
+            'reg',
+            (object, ),
+            {'records': {
+                'plone.resources.development': type(
+                    'val',
+                    (object, ),
+                    {'value': True}
+                )()
+            }}
+        )()
+        self.assertTrue(scripts.development)
+
+        scripts.update()
+        results = scripts.scripts()
+        self.assertEqual(
+            filter(lambda it: 'foo' in it['src'], results),
+            []
+        )
+
+    @mock.patch.object(
+        ScriptsView,
+        'get_resources',
+        new=lambda self: {'foo': DummyResource('++resource++foo.js')}
+    )
+    @mock.patch.object(
+        StylesView,
+        'get_resources',
+        new=lambda self: {'foo': DummyResource('++resource++foo.css')}
+    )
+    def test_add_resource_on_request_with_subrequest(self):
+        """Check, if a resource added at a main request is picked up from a
+        subrequest for creating the header scripts section.
+        """
+        req = self.layer['request']
+
+        # create a subrequest.
+        subreq = req.clone()
+        subreq['PARENT_REQUEST'] = req
+
+        # add a resource to main request
+        add_resource_on_request(req, 'foo')
+
+        scripts = ScriptsView(self.layer['portal'], subreq, None)
+        scripts.update()
+        results = scripts.scripts()
+        self.assertEqual(
+            results[-1],
+            {
+                'src': 'http://nohost/plone/++resource++foo.js',
+                'conditionalcomment': '',
+                'bundle': 'none'
+            }
+        )
+
+        styles = StylesView(self.layer['portal'], subreq, None)
+        styles.update()
+        results = styles.styles()
+        self.assertEqual(
+            filter(lambda it: 'foo' in it['src'], results)[0],
+            {
+                'src': 'http://nohost/plone/++resource++foo.css',
+                'conditionalcomment': '',
+                'rel': 'stylesheet',
+                'bundle': 'none',
+            }
+        )